Absolute Java Chapter14 Project 5
Implemented a Generic class with a type parameter.
Simulates a random draw of an item from this class.
Implemented a generic add method to add an item to an underlying
container of List<T> type.
Implemented an isEmpty method to see if the container is empty.
Implemented a draw method to randomly draw an item from the container.
Implemented a method to remove all items from the container.